pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "beam-mas"
version = "0.1.0"
description = "BEAM: Budget-Efficient Agent Management - Token-efficient multi-agent inference optimization"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.9"
authors = [
    {name = "BEAM Team"}
]
keywords = [
    "multi-agent",
    "llm",
    "optimization",
    "token-efficiency",
    "langchain",
    "langgraph"
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "torch>=2.0.0",
    "numpy>=1.21.0",
    "shortuuid>=1.0.0",
    "openai>=1.0.0",
    "tenacity>=8.0.0",
    "async-timeout>=4.0.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
langchain = [
    "langchain>=0.1.0",
    "langchain-core>=0.1.0",
    "langchain-openai>=0.0.5",
]
langgraph = [
    "langgraph>=0.0.20",
]
bayesian = [
    "scikit-learn>=1.0.0",
    "pyro-ppl>=1.8.0",
]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "mypy>=1.0.0",
]
all = [
    "beam-mas[langchain,langgraph,bayesian,dev]",
]

[project.urls]
Homepage = "https://github.com/beam-mas/beam"
Documentation = "https://github.com/beam-mas/beam#readme"
Repository = "https://github.com/beam-mas/beam"

[tool.setuptools.packages.find]
where = ["."]
include = ["beam*"]

[tool.black]
line-length = 100
target-version = ["py39", "py310", "py311"]

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
